refactoring of code

+++ src/services/apiCabins.js
@@ -11,56 +11,51 @@ export const getCabins = async () => {
   return data;
 };
 
-export const createEditCabin = async (newCabin, id) => {
+export async function createEditCabin(newCabin, id) {
   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-
   const imagePath = hasImagePath
     ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  //  create cabin
-
+  // 1. Create/edit cabin
   let query = supabase.from("cabins");
 
-  if (!id) {
-    query = query.insert([{ ...newCabin, image: imagePath }]);
-  }
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  // edit cabin
-
-  if (id) {
-    query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
-  }
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
   const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error);
-    throw new Error("Cabins could not be created");
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
 
-  // upload a file
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // delete cabin if error in uploading rnae
-
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-    console.log(storageError);
+    console.error(storageError);
     throw new Error(
       "Cabin image could not be uploaded and the cabin was not created"
     );
   }
 
   return data;
-};
+}
 
 export const deleteCabin = async (id) => {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
